chore!: re-do graphql schema snapshot testing
In the previous implementation we had the Jest test file in the package. This worked while we were
running tests in silverback-mono. But Jest always ignores test files which path includes
`node_modules`. So now the test file should be located in the tested project.
