--- packages/npm/@amazeelabs/graphql-schema-test/jest.config.run.js
@@ -1,3 +1,0 @@
-module.exports = {
-  testMatch: ['**/**'],
-};
