+++ packages/npm/@amazeelabs/graphql-schema-test/package.json
@@ -4,6 +4,8 @@
   "description": "Snapshot testing for GraphQL schemas.",
   "author": "Amazee Labs",
   "license": "MIT",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "private": false,
   "publishConfig": {
     "access": "public"
@@ -16,7 +18,6 @@
   },
   "dependencies": {
     "axios": "^0.24.0",
-    "jest": "^27",
     "jest-specific-snapshot": "^5.0.0",
     "lodash": "^4.17.21",
     "zx": "^4.2.0"
@@ -28,19 +29,17 @@
     "@types/jest": "^27",
     "@types/jest-specific-snapshot": "^0.5.5",
     "eslint": "^8",
+    "jest": "^27",
     "prettier": "^2",
     "typescript": "^4.5.4"
   },
   "scripts": {
-    "prepare": "rm -rf dist && tsc && chmod +x dist/graphql-schema-test.js",
+    "prepare": "rm -rf dist && tsc",
     "precommit": "lint-staged",
     "test:static": "tsc --noEmit && eslint \"**/*.{ts,tsx,js,jsx}\" --ignore-path=\"./.gitignore\" --fix",
     "test:unit": "jest --passWithNoTests",
     "test:integration": "exit 0",
     "test:watch": "jest --watch",
     "test": "yarn test:static && yarn test:unit && yarn test:integration"
-  },
-  "bin": {
-    "graphql-schema-test": "dist/graphql-schema-test.js"
   }
 }
