+++ packages/npm/@amazeelabs/graphql-schema-test/src/index.ts
@@ -1,0 +1,153 @@
+import axios from 'axios';
+import _ from 'lodash';
+import { fs } from 'zx';
+
+require('jest-specific-snapshot');
+const addSerializer = require('jest-specific-snapshot').addSerializer;
+
+export const listFiles = (
+  queriesDirPath: string,
+): Array<
+  [
+    /** Path relative to queriesDirPath. */
+    string,
+    /** Absolute path. */
+    string,
+  ]
+> => {
+  return walk(queriesDirPath).map((fullPath) => [
+    fullPath.substring(queriesDirPath.length + 1),
+    fullPath,
+  ]);
+};
+
+export const createExecutor = (
+  graphqlConfigPath: string,
+  serializer?: (responses: object) => object,
+): ((queryPath: string) => Promise<void>) => {
+  addSerializer({
+    serialize(responses: object) {
+      const processed = processData(responses);
+      if (serializer) {
+        serializer(processed);
+      }
+      return JSON.stringify(processed, null, 2);
+    },
+    test: () => true,
+  });
+
+  const endpoints: Record<
+    string,
+    { url: string; headers: Record<string, string> }
+  > = JSON.parse(fs.readFileSync(graphqlConfigPath, 'utf8')).extensions
+    .endpoints;
+
+  return async (queryPath: string) => {
+    const results = Object.fromEntries(
+      await Promise.all(
+        Object.entries(endpoints).map(async ([name, endpoint]) => {
+          const res = await axios.post(
+            endpoint.url,
+            { query: fs.readFileSync(queryPath).toString() },
+            {
+              headers: endpoint.headers,
+              validateStatus: null,
+            },
+          );
+          return [
+            name,
+            {
+              code: res.status,
+              data: res.data,
+            },
+          ];
+        }),
+      ),
+    );
+    expect(results).toMatchSpecificSnapshot(
+      // We can't use `.snap` extension as it conflicts with Jest's default
+      // snapshots.
+      `${queryPath}.snapshot`,
+    );
+  };
+};
+
+function processData(responses: any) {
+  let processed = responses;
+
+  // Replace Drupal's numeric IDs with "[numeric]" placeholder.
+  const processIds = (value: any) => {
+    return _.transform(value, (result, value, key) => {
+      if (
+        key === 'id' &&
+        _.isString(value) &&
+        value.match(/^[0-9]+(:[^:]+)?$/)
+      ) {
+        // @ts-ignore
+        result[key] = value.replace(/^([0-9]+)(:[^:]+)?$/, '[numeric]$2');
+      } else if (
+        key === 'drupalId' &&
+        _.isString(value) &&
+        value.match(/^[0-9]+$/)
+      ) {
+        // @ts-ignore
+        result[key] = '[numeric]';
+      } else if (
+        key === 'path' &&
+        _.isString(value) &&
+        value.match(/^.*\/[0-9]+$/)
+      ) {
+        // @ts-ignore
+        result[key] = value.replace(/^(.*\/)[0-9]+$/, '$1[numeric]');
+      } else {
+        if (_.isArray(value) || _.isObject(value)) {
+          // @ts-ignore
+          result[key] = processIds(value);
+        } else {
+          // @ts-ignore
+          result[key] = value;
+        }
+      }
+    });
+  };
+  processed = processIds(processed);
+
+  // Get rid of "locations" and "extensions" in "errors".
+  Object.keys(processed).forEach((key) => {
+    if (
+      _.has(processed[key], 'data.errors') &&
+      _.isArray(processed[key].data.errors)
+    ) {
+      for (const value of processed[key].data.errors) {
+        if (_.has(value, 'locations')) {
+          delete value.locations;
+        }
+        if (_.has(value, 'extensions')) {
+          delete value.extensions;
+        }
+      }
+    }
+  });
+
+  return processed;
+}
+
+// Can't use zx.glob because it's async and tests must be defined synchronously.
+// Function from https://stackoverflow.com/a/16684530/580371
+function walk(dir: string) {
+  let results: Array<string> = [];
+  const list = fs.readdirSync(dir);
+  list.forEach(function (file) {
+    file = dir + '/' + file;
+    const stat = fs.statSync(file);
+    if (stat && stat.isDirectory()) {
+      results = results.concat(walk(file));
+    } else {
+      if (file.endsWith('.gql')) {
+        // List only .gql files.
+        results.push(file);
+      }
+    }
+  });
+  return results;
+}
